Clean default url shorter route and build own metadata scraper  tool.
Main reason of this commit is, the main default link shorter api was too bloated for almost anything and for that reason it was resulting with a 500 error whenever someone tries to short a url. So to fix that I cleaned up that api and wrote my own metadata scraper library.